Moved Controller functions from View into Controller
